[Chore #161525972] add category role authentication

+++ app/api/v2/views/category_endpoint.py
@@ -2,8 +2,9 @@ from flask_restful import Resource, reqparse
 import psycopg2
 from flask import jsonify
 from ..models.category_model import Categories
+from ..models.user_model import Users
 from ..models import db_connection
-from flask_jwt_extended import (create_access_token, jwt_required, get_jwt_claims)
+from flask_jwt_extended import (create_access_token, jwt_required, get_jwt_claims, get_jwt_identity)
 
 #passing incoming data into post requests
 parser = reqparse.RequestParser()
@@ -12,17 +13,24 @@ parser.add_argument('name', help = 'This field cannot be blank', required = True
 class Category(Resource):
     @jwt_required
     def post(self):
-        """Post new category"""
+        """Post new category: only by the admin"""
         connection = db_connection()
         cursor = connection.cursor()
 
+        role = Users().get_user_role()
+
         data = parser.parse_args()
         name = data['name']
+
+        if role[0] != "admin":
+            return {
+                "message" : "Access not allowed"
+            },403
         
         if name.isalpha() == False:
             return{
                 'message' : 'Invalid category name'
-            }
+            },400
         else:
             try:
                 new_category = Categories()
@@ -31,10 +39,10 @@ class Category(Resource):
                 connection.commit()
                 
                 return {
-                        'message': 'Category created successfully',
+                        'message': 'Category created successfully'
                     },201
             except:
-                return {'message' : 'Category already exist'}
+                return {'message' : 'Category already exist'},409
     @jwt_required
     def get(self):
         """Get all Categories"""
@@ -69,7 +77,7 @@ class SingleCategory(Resource):
         data = cursor.fetchone()
 
         if data is None:
-            return {'message' : 'Category not Found'}
+            return {'message' : 'Category not Found'},404
 
         return {
             'message' : 'success',
@@ -78,10 +86,17 @@ class SingleCategory(Resource):
 
     @jwt_required
     def put(self, category_id):
-        """Modify one Category"""
+        """Modify one Category: only by the admin"""
 
         connection = db_connection()
         cursor = connection.cursor()
+
+        role = Users().get_user_role()
+
+        if role[0] != "admin":
+            return {
+                "message" : "Access not allowed"
+            },403
 
         data = parser.parse_args()
         name = data['name']
@@ -97,15 +112,21 @@ class SingleCategory(Resource):
                 },200
                 
         except:
-            return {'message': 'Category already exist'}
+            return {'message': 'Category already exist'},409
 
     @jwt_required
     def delete(self, category_id):
-        """delete one category"""
+        """delete one category: only by the admin"""
 
         connection = db_connection()
         cursor = connection.cursor()
 
+        role = Users().get_user_role()
+
+        if role[0] != "admin":
+            return {
+                "message" : "Access not allowed"
+            },403        
         try:
             category = Categories()
             sql = category.delete_category()
@@ -117,5 +138,5 @@ class SingleCategory(Resource):
               },200
 
         except:
-            return {'message' : 'Category not found'}
+            return {'message' : 'Category not found'},404
